added validation if wallet id is not a number, testing still not done

+++ server/controller/walletController.js
@@ -1,9 +1,10 @@
-
 const { Wallet } = require("../models/index");
 class WalletController {
   static async deleteWallet(req, res, next) {
     try {
       const { walletId } = req.params;
+      if (Number(walletId) !== "number") throw { name: "not a number" };
+      
       const deletedWallet = await Wallet.destroy({
         where: {
           id: walletId,
@@ -17,6 +18,8 @@ class WalletController {
         throw { name: "Data not found" };
       }
     } catch (error) {
+      if (error.name === "not a number")
+        res.status(404).json({ message: "Wallet ID is not a number" });
       next(error);
     }
   }
@@ -25,8 +28,11 @@ class WalletController {
     try {
       const { name, totalAmount } = req.body;
       const { walletId } = req.params;
+      if (Number(walletId) !== "number")
+        throw { name: "WalletId is not a number" };
+
       const updatedWallet = await Wallet.update(
-        { name, totalAmount },
+        { name, totalAmount: Number(totalAmount) },
         {
           where: {
             id: walletId,
@@ -41,7 +47,9 @@ class WalletController {
         throw { name: "Data not found" };
       }
     } catch (error) {
-        next(error)
+      if (error.name === "not a number")
+        res.status(404).json({ message: "Wallet ID is not a number" });
+      next(error);
     }
   }
 }
